+++ src/management-system/src/backend/server/iam/middleware/authorization.ts
@@ -1,6 +1,6 @@
 import { config } from '../utils/config.js';
 import { PackedRulesForUser, adminRules, rulesForUser } from '../authorization/caslRules';
-import Ability from '../authorization/abilityHelper';
+import Ability from '../../../../../../management-system-v2/lib/ability/abilityHelper';
 
 import Redis from 'ioredis';
