+++ src/management-system/src/backend/server/iam/rest-api/role-mappings.js
@@ -6,7 +6,7 @@ import {
   getRoleMappingByUserId,
 } from '../../../shared-electron-server/data/iam/role-mappings.js';
 import { deleteRulesForUsers, isAllowed } from '../middleware/authorization';
-import Ability from '../authorization/abilityHelper';
+import Ability from '../../../../../../management-system-v2/lib/ability/abilityHelper';
 import { toCaslResource } from '../authorization/caslRules';
 
 const roleMappingsRouter = express.Router();
