+++ src/management-system/src/backend/server/iam/rest-api/shares.js
@@ -10,7 +10,7 @@ import { abilityCacheDeleteAll, isAllowed } from '../middleware/authorization';
 import { validateShare } from '../middleware/inputValidations.js';
 import jwt from 'jsonwebtoken';
 import { toCaslResource } from '../authorization/caslRules';
-import Ability from '../authorization/abilityHelper';
+import Ability from '../../../../../../management-system-v2/lib/ability/abilityHelper';
 
 const sharesRouter = express.Router();
